refactor: move postgres migrate commands
